feat(data-table): extract a stateless table component
