+++ src/data-table/types.js
@@ -15,14 +15,16 @@ export type SortDirectionsT =
   | typeof SORT_DIRECTIONS.DESC
   | null;
 
+export type ColumnsT =
+  | typeof COLUMNS.BOOLEAN
+  | typeof COLUMNS.CATEGORICAL
+  | typeof COLUMNS.CUSTOM
+  | typeof COLUMNS.NUMERICAL
+  | typeof COLUMNS.STRING;
+
 // eslint-disable-next-line flowtype/no-weak-types
 export type ColumnT<ValueT = any, FilterParamsT = any> = {|
-  kind:
-    | typeof COLUMNS.BOOLEAN
-    | typeof COLUMNS.CATEGORICAL
-    | typeof COLUMNS.CUSTOM
-    | typeof COLUMNS.NUMERICAL
-    | typeof COLUMNS.STRING,
+  kind: ColumnsT,
   title: string,
   sortable: boolean,
   filterable: boolean,
@@ -65,10 +67,23 @@ export type RowActionT = {|
   renderIcon: React.ComponentType<{|size: number|}>,
 |};
 
-export type Props = {|
+export type StatefulDataTablePropsT = {|
   batchActions?: BatchActionT[],
   columns: ColumnT<>[],
   onSelectionChange?: (RowT[]) => mixed,
-  rowActions?: RowActionT[],
   rows: RowT[],
+  rowActions?: RowActionT[],
+|};
+
+export type DataTablePropsT = {|
+  ...StatefulDataTablePropsT,
+  filters: Map<string, mixed>,
+  onSelectMany: (RowT[]) => void,
+  onSelectNone: () => void,
+  onSelectOne: RowT => void,
+  onSort: number => void,
+  selectedRowIds: Set<string | number>,
+  sortIndex: number,
+  sortDirection: SortDirectionsT,
+  textQuery: string,
 |};
